<Feat> : 미션 인증 사진 조회 기능 추가 + (좋아요, 신고)

+++ project/Myo_jib_sa/app/src/main/java/com/example/myo_jib_sa/community/Retrofit/missionCert/missionResponse.kt
@@ -19,8 +19,8 @@ data class Rank(
     val userName:String
 )
 data class MissionProofImages(
-    val imageId:Int,
-    val userId:Int,
+    val imageId:Long,
+    val userId:Long,
     val filePath:String
 )
